+++ CHANGELOG.md
@@ -26,6 +26,8 @@
    `Quantity`.
  * [#7](https://github.com/iliekturtles/uom/issues/7) `Volume` quantity added. Additional units
    still need to be added.
+ * [#8](https://github.com/iliekturtles/uom/issues/8) `Force` quantity added. Additional units
+   still need to be added.
 
 ## [v0.14.0] — 2017-05-30